process list done
