+++ Makefile
@@ -12,6 +12,7 @@ SRCS	= \
 	srcs/create_token_util.c\
 	srcs/signal_handler.c\
 	srcs/create_process_list.c\
+	srcs/create_process_list_util.c\
 	srcs/utils/free_string_array.c\
 	srcs/utils/add_slash.c\
 	srcs/utils/remove_quotes.c\
