+++ srcs/create_process_list.c
@@ -6,40 +6,38 @@
 /*   By: yohatana <yohatana@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/23 13:03:04 by yohatana          #+#    #+#             */
-/*   Updated: 2025/03/23 17:22:30 by yohatana         ###   ########.fr       */
+/*   Updated: 2025/03/24 20:31:47 by yohatana         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include"../includes/minishell.h"
 
-static bool			add_to_cmd(t_proc_list **list, char *word);
-static bool			add_space(t_proc_list *curr);
+static bool			add_to_cmd(t_proc_list **list, \
+								char *word, \
+								bool new_proc_flg);
 static t_proc_list	*create_node(char *word);
 static t_proc		*create_proc(char *word);
 static bool			add_proc_list(t_proc_list **list, t_proc_list *new);
-t_proc_list			*get_last_proc(t_proc_list **list);
 
 t_proc_list	*create_process_list(t_token **head)
 {
 	t_proc_list	*list;
 	t_token		*curr;
 	bool		err_flg;
+	bool		new_proc_flg;
 
 	list = NULL;
 	err_flg = false;
 	curr = *head;
+	new_proc_flg = true;
 	while (curr)
 	{
 		if (ft_strcmp(curr->word, "|") == 0)
-		{
-			printf("this is pipe\n");
-			err_flg = add_proc_list(&list, create_node(NULL));
-			curr = curr->next;
-		}
+			new_proc_flg = true;
 		else
 		{
-			printf("add to cmd\n");
-			err_flg = add_to_cmd(&list, curr->word);
+			err_flg = add_to_cmd(&list, curr->word, new_proc_flg);
+			new_proc_flg = false;
 		}
 		if (err_flg)
 			break ;
@@ -80,7 +78,7 @@ static t_proc	*create_proc(char *word)
 	return (proc);
 }
 
-static bool	add_to_cmd(t_proc_list **list, char *word)
+static bool	add_to_cmd(t_proc_list **list, char *word, bool new_proc_flg)
 {
 	t_proc_list	*curr;
 	char		*temp;
@@ -88,19 +86,13 @@ static bool	add_to_cmd(t_proc_list **list, char *word)
 
 	err_flg = false;
 	curr = get_last_proc(list);
-	if (*list == NULL || curr->proc->cmd == NULL)
-	{
-		printf("new create node\n");
+	if (new_proc_flg)
 		err_flg = add_proc_list(list, create_node(word));
-	}
 	else
 	{
-		// curr = get_last_proc(list);
+		curr = get_last_proc(list);
 		if (curr->proc->cmd)
-		{
-			printf("add cmd\n");
 			err_flg = add_space(curr);
-		}
 		if (err_flg)
 			return (true);
 		temp = ft_strjoin(curr->proc->cmd, word);
@@ -134,27 +126,3 @@ static bool	add_proc_list(t_proc_list **list, t_proc_list *new)
 	return (false);
 }
 
-static bool	add_space(t_proc_list *curr)
-{
-	char	*temp;
-
-	temp = ft_strjoin(curr->proc->cmd, " ");
-	if (!temp)
-		return (true);
-	curr->proc->cmd = temp;
-	return (false);
-}
-
-t_proc_list	*get_last_proc(t_proc_list **list)
-{
-	t_proc_list	*temp;
-
-	if (*list == NULL)
-		return (*list);
-	temp = *list;
-	while (temp->next)
-	{
-		temp = temp->next;
-	}
-	return (temp);
-}
